+++ src/test/java/student_aleksandrs_korsaks/ak_lesson_11/level_5_6/fraud_detector_v2/FraudDetectorTest.java
@@ -1,4 +1,4 @@
-package student_aleksandrs_korsaks.ak_lesson_11.level_5.fraud_detector_v2;
+package student_aleksandrs_korsaks.ak_lesson_11.level_5_6.fraud_detector_v2;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static student_aleksandrs_korsaks.ak_lesson_11.level_5.fraud_detector_v2.FraudRule2.TRX_AMOUNT_LIMIT;
-import static student_aleksandrs_korsaks.ak_lesson_11.level_5.fraud_detector_v2.FraudRule5.TRX_AMOUNT_LIMIT_GERMANY;
+import static student_aleksandrs_korsaks.ak_lesson_11.level_5_6.fraud_detector_v2.FraudRule2.TRX_AMOUNT_LIMIT;
+import static student_aleksandrs_korsaks.ak_lesson_11.level_5_6.fraud_detector_v2.FraudRule5.TRX_AMOUNT_LIMIT_GERMANY;
 
 class FraudDetectorTest {
