+++ src/test/java/student_aleksandrs_korsaks/ak_lesson_11/level_5_6/fraud_detector_v2/FraudRule5Test.java
@@ -1,9 +1,9 @@
-package student_aleksandrs_korsaks.ak_lesson_11.level_5.fraud_detector_v2;
+package student_aleksandrs_korsaks.ak_lesson_11.level_5_6.fraud_detector_v2;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static student_aleksandrs_korsaks.ak_lesson_11.level_5.fraud_detector_v2.FraudRule5.TRX_AMOUNT_LIMIT_GERMANY;
+import static student_aleksandrs_korsaks.ak_lesson_11.level_5_6.fraud_detector_v2.FraudRule5.TRX_AMOUNT_LIMIT_GERMANY;
 
 class FraudRule5Test {
